+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/def/facade/EntityTestColumnRenameFacade.java
@@ -1,0 +1,11 @@
+package org.fugerit.java.daogen.sample.def.facade;
+
+/**
+ * EntityTestColumnRenameFacade, version : 1.0.0
+ *
+ * author: fugerit
+ */
+public interface EntityTestColumnRenameFacade extends EntityTestColumnRenameFacadeHelper {
+
+	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten) [do not delete this line]
+}