+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/rest/load/LoadTestColumnRename.java
@@ -1,0 +1,26 @@
+package org.fugerit.java.daogen.sample.impl.rest.load;
+
+import javax.ws.rs.Path;
+
+// custom import start ( code above here will be overwritten )
+// custom import end ( code below here will be overwritten )
+
+/**
+ * LoadTestColumnRename, version : 1.0.0
+ *
+ * author: fugerit
+ *
+ * warning!: auto generated object, insert custom code only between comments :
+ * // custom code start ( code above here will be overwritten )
+ * // custom code end ( code below here will be overwritten )
+ */
+@Path("/testcolumnrename/load")
+public class LoadTestColumnRename extends LoadTestColumnRenameHelper {
+
+	// custom code start ( code above here will be overwritten )
+	// custom code end ( code below here will be overwritten )
+
+	private static final long serialVersionUID = 104821631507L;
+
+	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten) [do not delete this line]
+}